Refactor simulation service commands and results
- Removed obsolete command and result classes related to simulation creation, deletion, and update.
- Introduced new packages for createSimulation, deleteSimulation, detailSimulation, and historySimulation to better organize the code.
- Updated imports across the application to reflect the new package structure.
- Ensured that all related service classes and interfaces are aligned with the new command and result DTOs.
