--- src/main/java/com/renewsim/backend/simulation_service/application/command/DeleteSimulationCommand.java
@@ -1,4 +1,0 @@
-package com.renewsim.backend.simulation_service.application.command;
-
-public record DeleteSimulationCommand(Long id) {
-}
