+++ src/main/java/com/renewsim/backend/simulation_service/application/deleteSimulation/DeleteAllSimulationsByUserCommand.java
@@ -1,4 +1,4 @@
-package com.renewsim.backend.simulation_service.application.command;
+package com.renewsim.backend.simulation_service.application.deleteSimulation;
 
 public record DeleteAllSimulationsByUserCommand(
       String username
