+++ src/main/java/com/renewsim/backend/simulation_service/application/deleteSimulation/DeleteAllSimulationsByUserUseCase.java
@@ -1,0 +1,7 @@
+package com.renewsim.backend.simulation_service.application.deleteSimulation;
+
+public interface DeleteAllSimulationsByUserUseCase {
+    
+    void deleteAllByUser(DeleteAllSimulationsByUserCommand command);
+
+}
