+++ src/main/java/com/renewsim/backend/simulation_service/application/deleteSimulation/DeleteSimulationCommand.java
@@ -1,0 +1,4 @@
+package com.renewsim.backend.simulation_service.application.deleteSimulation;
+
+public record DeleteSimulationCommand(Long id) {
+}
