+++ src/main/java/com/renewsim/backend/simulation_service/application/historySimulation/GetUserSimulationHistoryUseCase.java
@@ -1,0 +1,9 @@
+package com.renewsim.backend.simulation_service.application.historySimulation;
+
+
+import java.util.List;
+
+public interface GetUserSimulationHistoryUseCase {
+    List<SimulationHistoryResultDTO> getUserHistory(String username);
+}
+
