--- src/main/java/com/renewsim/backend/simulation_service/application/port/in/DeleteAllSimulationsByUserUseCase.java
@@ -1,9 +1,0 @@
-package com.renewsim.backend.simulation_service.application.port.in;
-
-import com.renewsim.backend.simulation_service.application.command.DeleteAllSimulationsByUserCommand;
-
-public interface DeleteAllSimulationsByUserUseCase {
-    
-    void deleteAllByUser(DeleteAllSimulationsByUserCommand command);
-
-}
