--- src/main/java/com/renewsim/backend/simulation_service/application/port/in/DeleteSimulationUseCase.java
@@ -1,11 +1,0 @@
-package com.renewsim.backend.simulation_service.application.port.in;
-
-import com.renewsim.backend.simulation_service.application.command.DeleteSimulationCommand;
-import com.renewsim.backend.simulation_service.application.result.SimulationDeletionResultDTO;
-
-/**
- * Use case for deleting a Simulation by its ID.
- */
-public interface DeleteSimulationUseCase {
-    SimulationDeletionResultDTO deleteSimulation(DeleteSimulationCommand command);
-}
