+++ src/main/java/com/renewsim/backend/simulation_service/application/port/in/GetSimulationUseCase.java
@@ -1,7 +1,7 @@
 package com.renewsim.backend.simulation_service.application.port.in;
 
 import com.renewsim.backend.simulation_service.application.command.GetSimulationByIdCommand;
-import com.renewsim.backend.simulation_service.application.result.SimulationDetailResultDTO;
+import com.renewsim.backend.simulation_service.application.detailSimulation.SimulationDetailResultDTO;
 
 /**
  * Use case for retrieving a Simulation by ID.
