+++ src/main/java/com/renewsim/backend/simulation_service/application/service/SimulationApplicationService.java
@@ -9,6 +9,15 @@ import org.springframework.transaction.annotation.Transactional;
 import com.renewsim.backend.simulation_service.application.port.in.*;
 import com.renewsim.backend.simulation_service.application.port.out.*;
 import com.renewsim.backend.simulation_service.application.command.*;
+import com.renewsim.backend.simulation_service.application.createSimulation.CreateSimulationCommand;
+import com.renewsim.backend.simulation_service.application.createSimulation.CreateSimulationUseCase;
+import com.renewsim.backend.simulation_service.application.createSimulation.SimulationCreationResultDTO;
+import com.renewsim.backend.simulation_service.application.deleteSimulation.DeleteSimulationCommand;
+import com.renewsim.backend.simulation_service.application.deleteSimulation.DeleteSimulationUseCase;
+import com.renewsim.backend.simulation_service.application.deleteSimulation.SimulationDeletionResultDTO;
+import com.renewsim.backend.simulation_service.application.detailSimulation.SimulationDetailResultDTO;
+import com.renewsim.backend.simulation_service.application.historySimulation.GetUserSimulationHistoryUseCase;
+import com.renewsim.backend.simulation_service.application.historySimulation.SimulationHistoryResultDTO;
 import com.renewsim.backend.simulation_service.application.result.*;
 import com.renewsim.backend.simulation_service.domain.factory.SimulationFactory;
 import com.renewsim.backend.simulation_service.domain.model.Simulation;
@@ -20,6 +29,9 @@ import com.renewsim.backend.simulation_service.domain.model.vo.EnergyOutput;
 import com.renewsim.backend.shared.exception.ConflictException;
 import com.renewsim.backend.simulation_service.domain.model.vo.ProjectSize;
 import com.renewsim.backend.simulation_service.application.result.SimulationRecommendationResultDTO;
+import com.renewsim.backend.simulation_service.application.updateSimulation.SimulationUpdateResultDTO;
+import com.renewsim.backend.simulation_service.application.updateSimulation.UpdateSimulationCommand;
+import com.renewsim.backend.simulation_service.application.updateSimulation.UpdateSimulationUseCase;
 
 import java.time.LocalDateTime;
 import java.util.List;
