+++ src/main/java/com/renewsim/backend/simulation_service/application/service/SimulationOrchestratorService.java
@@ -1,6 +1,6 @@
 package com.renewsim.backend.simulation_service.application.service;
 
-import com.renewsim.backend.simulation_service.application.command.CreateSimulationCommand;
+import com.renewsim.backend.simulation_service.application.createSimulation.CreateSimulationCommand;
 import com.renewsim.backend.simulation_service.application.result.SimulationRecommendationResultDTO;
 
 import lombok.RequiredArgsConstructor;
