+++ src/main/java/com/renewsim/backend/simulation_service/application/updateSimulation/UpdateSimulationUseCase.java
@@ -1,0 +1,8 @@
+package com.renewsim.backend.simulation_service.application.updateSimulation;
+
+/**
+ * Use case for updating an existing Simulation.
+ */
+public interface UpdateSimulationUseCase {
+    SimulationUpdateResultDTO updateSimulation(UpdateSimulationCommand command);
+}
