+++ src/main/java/com/renewsim/backend/simulation_service/web/controller/SimulationController.java
@@ -2,8 +2,21 @@ package com.renewsim.backend.simulation_service.web.controller;
 
 import com.renewsim.backend.auth_service.domain.AuthenticatedUser;
 import com.renewsim.backend.simulation_service.application.command.*;
+import com.renewsim.backend.simulation_service.application.createSimulation.CreateSimulationCommand;
+import com.renewsim.backend.simulation_service.application.createSimulation.CreateSimulationUseCase;
+import com.renewsim.backend.simulation_service.application.createSimulation.SimulationCreationResultDTO;
+import com.renewsim.backend.simulation_service.application.deleteSimulation.DeleteAllSimulationsByUserCommand;
+import com.renewsim.backend.simulation_service.application.deleteSimulation.DeleteAllSimulationsByUserUseCase;
+import com.renewsim.backend.simulation_service.application.deleteSimulation.DeleteSimulationCommand;
+import com.renewsim.backend.simulation_service.application.deleteSimulation.DeleteSimulationUseCase;
+import com.renewsim.backend.simulation_service.application.deleteSimulation.SimulationDeletionResultDTO;
+import com.renewsim.backend.simulation_service.application.detailSimulation.SimulationDetailResultDTO;
+import com.renewsim.backend.simulation_service.application.historySimulation.GetUserSimulationHistoryUseCase;
 import com.renewsim.backend.simulation_service.application.port.in.*;
 import com.renewsim.backend.simulation_service.application.result.*;
+import com.renewsim.backend.simulation_service.application.updateSimulation.SimulationUpdateResultDTO;
+import com.renewsim.backend.simulation_service.application.updateSimulation.UpdateSimulationCommand;
+import com.renewsim.backend.simulation_service.application.updateSimulation.UpdateSimulationUseCase;
 import com.renewsim.backend.simulation_service.domain.model.vo.EnergyType;
 import com.renewsim.backend.simulation_service.domain.model.vo.ResolvedLocation;
 import com.renewsim.backend.simulation_service.application.port.out.ClimateDataProviderPort;
