+++ src/test/java/com/renewsim/backend/simulation_service/web/controller/SimulationControllerTest.java
@@ -1,22 +1,22 @@
 package com.renewsim.backend.simulation_service.web.controller;
 
 import com.renewsim.backend.auth_service.domain.AuthenticatedUser;
-import com.renewsim.backend.simulation_service.application.command.CreateSimulationCommand;
 import com.renewsim.backend.simulation_service.application.command.GetSimulationByIdCommand;
-import com.renewsim.backend.simulation_service.application.port.in.CreateSimulationUseCase;
-import com.renewsim.backend.simulation_service.application.port.in.DeleteAllSimulationsByUserUseCase;
-import com.renewsim.backend.simulation_service.application.port.in.DeleteSimulationUseCase;
+import com.renewsim.backend.simulation_service.application.createSimulation.CreateSimulationCommand;
+import com.renewsim.backend.simulation_service.application.createSimulation.CreateSimulationUseCase;
+import com.renewsim.backend.simulation_service.application.createSimulation.SimulationCreationResultDTO;
+import com.renewsim.backend.simulation_service.application.deleteSimulation.DeleteAllSimulationsByUserUseCase;
+import com.renewsim.backend.simulation_service.application.deleteSimulation.DeleteSimulationUseCase;
+import com.renewsim.backend.simulation_service.application.detailSimulation.SimulationDetailResultDTO;
+import com.renewsim.backend.simulation_service.application.historySimulation.GetUserSimulationHistoryUseCase;
+import com.renewsim.backend.simulation_service.application.historySimulation.SimulationHistoryResultDTO;
 import com.renewsim.backend.simulation_service.application.port.in.GetSimulationUseCase;
 import com.renewsim.backend.simulation_service.application.port.in.GetSimulationDashboardSummaryUseCase;
-import com.renewsim.backend.simulation_service.application.port.in.GetUserSimulationHistoryUseCase;
-import com.renewsim.backend.simulation_service.application.port.in.UpdateSimulationUseCase;
 import com.renewsim.backend.simulation_service.application.port.out.ClimateDataProviderPort;
-import com.renewsim.backend.simulation_service.application.result.SimulationCreationResultDTO;
 import com.renewsim.backend.simulation_service.application.result.SimulationDashboardSummaryResult;
 import com.renewsim.backend.simulation_service.application.result.SimulationDashboardStatsResult;
 import com.renewsim.backend.simulation_service.application.result.SimulationDashboardEnergyBySourceResult;
-import com.renewsim.backend.simulation_service.application.result.SimulationDetailResultDTO;
-import com.renewsim.backend.simulation_service.application.result.SimulationHistoryResultDTO;
+import com.renewsim.backend.simulation_service.application.updateSimulation.UpdateSimulationUseCase;
 import com.renewsim.backend.simulation_service.domain.model.vo.ClimateData;
 import com.renewsim.backend.simulation_service.domain.model.vo.ResolvedLocation;
 import com.renewsim.backend.simulation_service.web.dto.CreateSimulationResponseDTO;
